Restaurant System
During my enrollment in a Data Structures course at university, I undertook the development of a restaurant system project. This initiative involved the utilization of Java programming language along with an array of advanced data structures, including queues, stacks, and trees. The primary aim was to demonstrate my adept understanding and proficient application of these intricate data structures and methodologies. This project not only served as a practical application of theoretical concepts learned in the course but also highlighted my ability to implement complex algorithms effectively.